Generate markdown for agentk and kas config
This change sets adds the `extract_generated_proto_markdown_doc`
target to `//pkg/kascfg` and `//pkg/agentcfg`.
They will generate markdown documentation from the proto files and
extract them from the build folder to the workspace in the appropriate
`/pkg/kascfg` / `/pkg/agentcfg` folder.

The adaption of the proto spec comments is out of scope of this MR on
purpose and I'll do this in a followup MR.

Issue: https://gitlab.com/gitlab-org/cluster-integration/gitlab-agent/-/issues/195