Verify correct Proxy API path prefix before authentication
This change set moves the verification to check that the Proxy URL
contains the correct prefix before authenticating the request.

The reason for this is that for
https://gitlab.com/gitlab-org/gitlab/-/issues/412163 a natural place to
implement usage metrics is the authenication. However, we don't
necessarily want these metrics in case the API URL is prefixed
incorrectly.

The "downside" of this change is that a user can learn about the
configured API prefix. However, I don't see this as an actual problem.
It would more be security-by-obscurity.