Report more fine-grained CI tunnel usage metrics
There are 6 new metrics introduced with this change set that can be
categorized into 2 categories: CI tunnel access via `ci_access` and the
other via `user_access`.

Each of the categories have 3 distinct metrics:

1. number of requests
1. unique number of users (by user id)
1. unique number of agents (by agent id)

The old metrics for total number of requests and unique users stay in
place for now, but could easily be removed in the future in case we
don't need them anymore.