Always run leader election
This changes the leader runner behavior which previously conditionally
depending on if there were any leader modules registered would start the
election or not. Now, it always starts the election.

However, in reality this doesn't happen, because e.g. the
`startboard_vulnerability` module is a leader-only module an is always
enabled.